Vacation History Completes

+++ EPMRepositories/Mockups/IUsersMockupRepository.cs
@@ -18,11 +18,26 @@ namespace EPMServices
 				Password = "Password"
 			};
 			_users.Add (sandro);
-			_users.Add (new User{ Id = 2, Name = "Leonel Messi", VacationDays = 6, Account = "lmessi", Password = "Password", Manager = sandro });
-			_users.Add (new User{ Id = 3, Name = "Xavi Hernandez", VacationDays = 6, Account = "xhernandez", Password = "Password", Manager = sandro });
+			_users.Add (new User {
+				Id = 2,
+				Name = "Leonel Messi",
+				VacationDays = 6,
+				Account = "lmessi",
+				Password = "Password",
+				Manager = sandro
+			});
+			_users.Add (new User {
+				Id = 3,
+				Name = "Xavi Hernandez",
+				VacationDays = 6,
+				Account = "xhernandez",
+				Password = "Password",
+				Manager = sandro
+			});
 		}
 
-		public User GetUserById (long Id){
+		public User GetUserById (Int32 Id)
+		{
 			foreach (var u in _users) {
 				if (u.Id.Equals (Id))
 					return u;
@@ -30,7 +45,8 @@ namespace EPMServices
 			return null;
 		}
 
-		public long IsLoginValid(string acc, string pass){
+		public long IsLoginValid (string acc, string pass)
+		{
 			foreach (var u in _users) {
 				if (u.Account.Equals (acc)) {
 					if (u.Password.Equals (pass))
@@ -42,7 +58,8 @@ namespace EPMServices
 			return -1;
 		}
 
-		public IEnumerable<User> users(){
+		public IEnumerable<User> users ()
+		{
 			return _users;
 		}
 	}
